requirements.txt
dlib==19.22.1
imutils==0.5.4
numpy==1.19.5
opencv_python==4.5.5.64
Pillow==9.1.0
PyQt5==5.15.6
scipy==1.6.2
torchvision==0.12.0
torch==1.10.2+cu102

